modified

+++ railway-rsv-service/src/main/java/com/web/rail/dtos/ScheduleNewTrainDTO.java
@@ -1,5 +1,5 @@
 package com.web.rail.dtos;
 
-public record ScheduleNewTrainDTO(int id,
+public record ScheduleNewTrainDTO(Long id,
                                   String dateTime) {
 }
